requirements.txt
google-generativeai==0.8.5
python-telegram-bot==22.4
python-dotenv==1.1.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
googleapis-common-protos==1.70.0
textbase
